Implement custom LFO node
 * supports resetting phase when playback starts and computing phase when starting at a point other than beat 0
 * move a bunch of wavetable fm synth code around to support pulling the oscillators out into the LFO node impl without having the imports
 * build AWP, instance, UI, and plumbing for the LFO AWP, handling converting old state into new state and making everything backwards compatible

+++ engine/wavetable/src/fm/effects/comb_filter.rs
@@ -1,9 +1,10 @@
-use dsp::circular_buffer::CircularBuffer;
+use dsp::{circular_buffer::CircularBuffer, FRAME_SIZE, SAMPLE_RATE};
+
+use crate::fm::param_source::ParamSource;
 
 use super::Effect;
-use crate::fm::{ParamSource, FRAME_SIZE, SAMPLE_RATE};
 
-pub const MAX_DELAY_SAMPLES: usize = SAMPLE_RATE * 4;
+pub const MAX_DELAY_SAMPLES: usize = SAMPLE_RATE as usize * 4;
 
 #[derive(Clone)]
 pub struct CombFilter {
